feat(shiki): expose highlighting utils (#1727)
