+++ src/module.ts
@@ -299,7 +299,7 @@ export default defineNuxtModule<ModuleOptions>({
       config.optimizeDeps = config.optimizeDeps || {}
       config.optimizeDeps.include = config.optimizeDeps.include || []
       config.optimizeDeps.include.push(
-        'html-tags'
+        'html-tags', 'slugify'
       )
     })
