+++ src/runtime/types.d.ts
@@ -501,15 +501,3 @@ export interface NavItem {
 
   [key: string]: any
 }
-
-// Highlight
-export interface HighlightParams {
-  code: string
-  lang: string
-  theme: Theme | Record<string, Theme>
-}
-
-export interface HighlightThemedToken {
-  content: string
-  color?: string | Record<string, string>
-}
